fix(ci): address rector and phpstan errors after rebase
- search_code.php: drop redundant (int) cast on $M + 1 ($M is a
  literal int, Rector flagged it).
- dynamic_finder_ajax.php, find_code_dynamic_ajax.php: wrap
  filter_input(INPUT_GET, X) ?? '' in (string) (...) since PHPStan
  types the result as string|false|null. The outer cast normalises
  to a definite string, fixing downstream explode, array offset,
  and main_code_set_search parameter errors.
- demographics.php: drop (int) $limit cast in get_document_by_catg().
  The function's only caller passes literal 3 and the default is 1,
  so the cast is redundant and was tripping the "Cannot cast mixed
  to int" baseline count for this file (4 expected, 5 found).

+++ interface/patient_file/encounter/search_code.php
@@ -87,7 +87,7 @@ if (isset($_POST["mode"]) && $_POST["mode"] == "search" && $_POST["text"] != "")
     " LIMIT ?" .
     "";
 
-    if ($res = sqlStatement($sql, [$pid, "%" . $_POST["text"] . "%", "%" . $_POST["text"] . "%", $code_types[$code_type]['id'], (int) ($M + 1)])) {
+    if ($res = sqlStatement($sql, [$pid, "%" . $_POST["text"] . "%", "%" . $_POST["text"] . "%", $code_types[$code_type]['id'], $M + 1])) {
         $result = [];
         for ($iter = 0; $row = sqlFetchArray($res); $iter++) {
             $result[$iter] = $row;
